This fix will allow spark to reuse exchange when table and pushed down filters are same
