+++ spark-bigquery-dsv2/spark-bigquery-dsv2-common/src/main/java/com/google/cloud/spark/bigquery/v2/context/BigQueryDataSourceReaderContext.java
@@ -322,6 +322,10 @@ public class BigQueryDataSourceReaderContext {
     return BigQueryUtil.friendlyTableName(tableId);
   }
 
+  public TableId getTableId() {
+    return tableId;
+  }
+
   public BigQueryRDDFactory getBigQueryRddFactory() {
     return this.bigQueryRDDFactory;
   }
